+++ pajbot/migration/migrate.py
@@ -2,7 +2,10 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, Any, List, Optional, Union
 
+import sys
 import importlib
+from importlib.abc import PathEntryFinder, MetaPathFinder
+from importlib.util import module_from_spec
 import logging
 import pkgutil
 import re
@@ -69,11 +72,17 @@ class Migration:
             if ispkg:
                 continue
 
-            if isinstance(importer, importlib.abc.PathEntryFinder):
-                maybe_module = importer.find_module(modname)
-                if maybe_module is None:
+            if isinstance(importer, PathEntryFinder):
+                module_spec = importer.find_spec(modname)
+                if module_spec is None:
                     continue
-                module = maybe_module.load_module(modname)
+                if module_spec.loader is None:
+                    # Module spec did not manage to make a loader, probably not a python file?
+                    continue
+                module = module_from_spec(module_spec)
+
+                module_spec.loader.exec_module(module)
+
                 id_in_module = getattr(module, "ID", None)
                 name_in_module = getattr(module, "NAME", None)
                 up_action = getattr(module, "up", None)
@@ -115,7 +124,7 @@ class Migration:
                 revision = Revision(id, name, up_action)
 
                 revisions.append(revision)
-            elif isinstance(importer, importlib.abc.MetaPathFinder):
+            elif isinstance(importer, MetaPathFinder):
                 log.info(f"Not sure how to handle metapathfinder {importer}, {type(importer)}")
 
         if len(revisions) <= 0: